refactor: 에러메시지를 상태코드를 포함한 enum으로 리팩토링하기 위해 유효성 검사 에러메시지를 따로 분리
- 분리해준 이유는 enum은 getMessage()를 하더라도 어노테이션 속성에 지정할 수 없기 때문에(컴파일 상수만 가능) 따로 분리함
