+++ src/main/java/com/peeerr/climbing/controller/FileController.java
@@ -1,8 +1,8 @@
 package com.peeerr.climbing.controller;
 
-import com.peeerr.climbing.exception.ErrorMessage;
 import com.peeerr.climbing.dto.common.ApiResponse;
 import com.peeerr.climbing.exception.ClimbingException;
+import com.peeerr.climbing.exception.ValidationErrorMessage;
 import com.peeerr.climbing.security.MemberPrincipal;
 import com.peeerr.climbing.service.FileService;
 import lombok.RequiredArgsConstructor;
@@ -34,7 +34,7 @@ public class FileController {
                                            @RequestParam List<MultipartFile> files,
                                            @AuthenticationPrincipal MemberPrincipal userDetails) {
         if (files == null || files.isEmpty() || files.stream().anyMatch(file -> file == null || file.isEmpty())) {
-            throw new ClimbingException(ErrorMessage.FILE_REQUIRED);
+            throw new ClimbingException(ValidationErrorMessage.FILE_REQUIRED);
         }
 
         Long loginId = userDetails.getMember().getId();
