+++ src/main/java/com/peeerr/climbing/dto/request/CommentEditRequest.java
@@ -7,7 +7,7 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-import static com.peeerr.climbing.exception.ErrorMessage.CONTENT_NOT_BLANK;
+import static com.peeerr.climbing.exception.ValidationErrorMessage.CONTENT_NOT_BLANK;
 
 @NoArgsConstructor
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
